requirements.txt
Flask>=3.0.0
numpy>=2.3.0
pandas>=2.3.0
pymongo>=4.5.0
scikit-learn>=1.5.0
gunicorn>=21.2.0
xgboost>=2.1.0
pymongo[srv]
python-dotenv>=1.0.1
requests>=2.32.0
